+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/ContainerRunningMonitor.kt
@@ -5,18 +5,18 @@ import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toConditionalValue
 import com.krillsson.sysapi.core.domain.system.SystemInfo
-import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
+import com.krillsson.sysapi.core.monitoring.MonitorInput
 import java.util.*
 
-class DockerContainerRunningMonitor(
+class ContainerRunningMonitor(
     override val id: UUID,
     override val config: MonitorConfig<MonitoredValue.ConditionalValue>
 ) : Monitor<MonitoredValue.ConditionalValue>() {
     override val type: Type = Type.CONTAINER_RUNNING
 
     companion object {
-        val selector: ContainerConditionalValueSelector = { containers, monitoredItemId ->
+        val selector: ContainerConditionalValueSelector = { containers, containerMetrics, monitoredItemId ->
             containers.filter {
                 it.id.equals(monitoredItemId, ignoreCase = true)
             }.map {
@@ -29,8 +29,8 @@ class DockerContainerRunningMonitor(
         }
     }
 
-    override fun selectValue(event: MetricQueryEvent): MonitoredValue.ConditionalValue? {
-        return selector(event.containers, config.monitoredItemId)
+    override fun selectValue(event: MonitorInput): MonitoredValue.ConditionalValue? {
+        return selector(event.containers, event.containerStats, config.monitoredItemId)
     }
 
     override fun maxValue(info: SystemInfo): MonitoredValue.ConditionalValue? {