+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/CpuTemperatureMonitor.kt
@@ -3,8 +3,8 @@ package com.krillsson.sysapi.core.monitoring.monitors
 import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.system.SystemInfo
-import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
+import com.krillsson.sysapi.core.monitoring.MonitorInput
 import java.util.*
 
 class CpuTemperatureMonitor(override val id: UUID, override val config: MonitorConfig<MonitoredValue.NumericalValue>) :
@@ -18,7 +18,7 @@ class CpuTemperatureMonitor(override val id: UUID, override val config: MonitorC
 
     override val type: Type = Type.CPU_TEMP
 
-    override fun selectValue(event: MetricQueryEvent): MonitoredValue.NumericalValue? =
+    override fun selectValue(event: MonitorInput): MonitoredValue.NumericalValue? =
         selector(event.load, null)
 
     override fun maxValue(info: SystemInfo): MonitoredValue.NumericalValue? {