+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/DiskWriteRateMonitor.kt
@@ -5,8 +5,8 @@ import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toNumericalValue
 import com.krillsson.sysapi.core.domain.system.SystemInfo
-import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
+import com.krillsson.sysapi.core.monitoring.MonitorInput
 import java.util.*
 
 class DiskWriteRateMonitor(override val id: UUID, override val config: MonitorConfig<MonitoredValue.NumericalValue>) :
@@ -24,7 +24,7 @@ class DiskWriteRateMonitor(override val id: UUID, override val config: MonitorCo
         }
     }
 
-    override fun selectValue(event: MetricQueryEvent): MonitoredValue.NumericalValue? =
+    override fun selectValue(event: MonitorInput): MonitoredValue.NumericalValue? =
         selector(event.load, config.monitoredItemId)
 
     override fun maxValue(info: SystemInfo): MonitoredValue.NumericalValue? {