+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/FileSystemSpaceMonitor.kt
@@ -6,8 +6,8 @@ import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toNumericalValue
 import com.krillsson.sysapi.core.domain.system.SystemInfo
-import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
+import com.krillsson.sysapi.core.monitoring.MonitorInput
 import java.util.*
 
 class FileSystemSpaceMonitor(override val id: UUID, override val config: MonitorConfig<MonitoredValue.NumericalValue>) :
@@ -28,7 +28,7 @@ class FileSystemSpaceMonitor(override val id: UUID, override val config: Monitor
         }
     }
 
-    override fun selectValue(event: MetricQueryEvent): MonitoredValue.NumericalValue? =
+    override fun selectValue(event: MonitorInput): MonitoredValue.NumericalValue? =
         selector(event.load, config.monitoredItemId)
 
     override fun maxValue(info: SystemInfo): MonitoredValue.NumericalValue? {