+++ server/src/main/java/com/krillsson/sysapi/core/monitoring/monitors/ProcessExistsMonitor.kt
@@ -4,8 +4,8 @@ import com.krillsson.sysapi.core.domain.monitor.MonitorConfig
 import com.krillsson.sysapi.core.domain.monitor.MonitoredValue
 import com.krillsson.sysapi.core.domain.monitor.toConditionalValue
 import com.krillsson.sysapi.core.domain.system.SystemInfo
-import com.krillsson.sysapi.core.monitoring.MetricQueryEvent
 import com.krillsson.sysapi.core.monitoring.Monitor
+import com.krillsson.sysapi.core.monitoring.MonitorInput
 import java.util.*
 
 class ProcessExistsMonitor(
@@ -22,7 +22,7 @@ class ProcessExistsMonitor(
 
     override val type: Type = Type.PROCESS_EXISTS
 
-    override fun selectValue(event: MetricQueryEvent): MonitoredValue.ConditionalValue? =
+    override fun selectValue(event: MonitorInput): MonitoredValue.ConditionalValue? =
         selector(event.load, config.monitoredItemId)
 
     override fun maxValue(info: SystemInfo): MonitoredValue.ConditionalValue {